+++ src/helpers/sls-utils/get-data.js
@@ -9,20 +9,20 @@ module.exports = (modules, meta, type) => {
 
 function getLayerData (modules, layers) {
   return {
-    layers: require('../main-handlers/layer/get-packages.js')(modules, layers)
+    layers: require('../../main-handlers/layer/get-packages.js')(modules, layers)
   }
 }
 
 function getLambdaData (modules, layers, functions) {
   return {
     ...getLayerData(modules, layers),
-    functions: require('../main-handlers/lambda/process-meta.js')(functions, layers)
+    functions: require('../../main-handlers/lambda/process-meta.js')(functions, layers)
   }
 }
 
 function getApiData (modules, layers, functions) {
   return {
     ...getLambdaData(modules, layers, functions),
-    endpoints: require('../main-handlers/api/get-endpoints.js')(functions)
+    endpoints: require('../../main-handlers/api/get-endpoints.js')(functions)
   }
 }
